Ajustes Exercícios 9 e 10

+++ caderno-de-exercicios/exercicio09/src/Aluno.java
@@ -1,5 +1,5 @@
 public class Aluno {
     String nome;
-    int matricula;
+    long matricula;
     boolean bolsista;
 }
